news_analyzer/requirements.txt
google-genai==1.56.0
psycopg[binary]==3.3.2
python-dotenv==1.2.1
tenacity==9.1.2
requests==2.32.5
beautifulsoup4==4.14.3
pydantic==2.12.5
numpy==2.4.0
pgvector==0.4.2
pytz==2024.2
